Layout export didn't properly reimport permissions
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@566 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-ejb/src/com/liferay/portal/service/impl/RoleLocalServiceImpl.java
@@ -115,9 +115,13 @@ public class RoleLocalServiceImpl implements RoleLocalService {
 
 		// Resources
 
-		ResourceLocalServiceUtil.deleteResource(
-			role.getCompanyId(), Role.class.getName(), Resource.TYPE_CLASS,
-			Resource.SCOPE_INDIVIDUAL, role.getPrimaryKey().toString());
+		if (Validator.isNull(role.getClassName()) &&
+			Validator.isNull(role.getClassPK())) {
+
+			ResourceLocalServiceUtil.deleteResource(
+				role.getCompanyId(), Role.class.getName(), Resource.TYPE_CLASS,
+				Resource.SCOPE_INDIVIDUAL, role.getPrimaryKey().toString());
+		}
 
 		// Role
